requirements.txt
Django
djangorestframework
qrcode[pil]
isort
flake8
pdfkit